feat(volcano, input): Start work on prototype input-/event-dispatcher
ATM it's not much. The idea is to maybe have a centralized class where
input-event will be sent to in order to be dispatched afterwards.

It should be possible to add/remove listeners/callbacks for all inputs
and have them be dispatched in a dynamic fashion.

I'm not really sure about the implementation right now and maybe this
is just a bad idea - I don't know.